Rewrite test to structured modules
