+++ ci-runtime-prepare.sh
@@ -7,13 +7,13 @@ mkdir -p ./runtimes/.test/helpers
 cp -R ./helpers/* ./runtimes/.test/helpers
 
 # Runtime base dockerfile
-cp -R ./runtimes/$RUNTIME/$RUNTIME.dockerfile ./runtimes/.test
+cp -R ./runtimes/$RUNTIME_FOLDER/$RUNTIME_FOLDER.dockerfile ./runtimes/.test
 
 # Runtime-specific files (most)
-cp -R ./runtimes/$RUNTIME/versions/latest/* ./runtimes/.test
+cp -R ./runtimes/$RUNTIME_FOLDER/versions/latest/* ./runtimes/.test
 
 # Version-specific files
-cp -R ./runtimes/$RUNTIME/versions/$VERSION/* ./runtimes/.test
+cp -R ./runtimes/$RUNTIME_FOLDER/versions/$VERSION_FOLDER/* ./runtimes/.test
 
 # Global Docker configuration
 cp ./base-before.dockerfile ./runtimes/.test/base-before.dockerfile
