+++ tests/serverless/Bun.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Serverless as TestsServerless;
+
+class Python extends TestsServerless
+{
+}
