+++ tests/serverless/Cpp.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Serverless as TestsServerless;
+
+class Cpp extends TestsServerless
+{
+}
