+++ tests/serverless/Dart.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Serverless as TestsServerless;
+
+class Dart extends TestsServerless
+{
+}
