+++ tests/serverless/Dotnet.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Serverless as TestsServerless;
+
+class Dotnet extends TestsServerless
+{
+}
