+++ tests/serverless/Java.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Serverless as TestsServerless;
+
+class Java extends TestsServerless
+{
+}
