+++ tests/serverless/Php.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Serverless as TestsServerless;
+
+class Php extends TestsServerless
+{
+}
