+++ tests/serverless/PythonMl.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Client;
+
+class PythonMl extends Python
+{
+    public function testTensorflowVersion(): void
+    {
+        $response = Client::execute(body: 'Hello', headers: ['x-action' => 'tensorflowVersion']);
+
+        self::assertEquals(200, $response['code']);
+        self::assertEquals("22", $response['body']);
+    }
+}
