+++ tests/serverless/Ruby.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Serverless as TestsServerless;
+
+class Ruby extends TestsServerless
+{
+}
