+++ tests/serverless/Swift.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Serverless;
+
+use Tests\Serverless as TestsServerless;
+
+class Swift extends TestsServerless
+{
+}
